pyproject.toml
[project]
name = "django-pytest-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "pytest>=8.3.2",
    "pytest-django>=4.9.0",
    "model-bakery>=1.19.5",
    "psycopg2-binary>=2.9.9",
    "django-pgtrigger>=4.12.2",
    "typing-extensions>=4.12.2",
    "pytest-sugar>=1.0.0",
    "django-filter>=24.3",
]

[tool.ruff]
line-length = 140

[tool.uv]
dev-dependencies = []

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.pytest_example.settings"
pythonpath = ". src"


